Move utils.js to helpers folder
